feat: extracted settings form component, and introduced a repair page
