+++ web/angular-wallet/proxy.config.json
@@ -1,6 +1,6 @@
 {
   "/burst": {
-    "target": "https://testnet-2.burst-alliance.org:6876",
+    "target": "https://testnet.burstcoin.network:6876",
     "changeOrigin": true,
     "secure": false,
     "logLevel": "info"
