+++ web/angular-wallet/src/app/app.theme.scss
@@ -81,7 +81,7 @@ $typography: mat-typography-config(
 @import "src/app/layout/components/chat-panel/chat-panel.theme";
 @import "src/app/layout/components/toolbar/toolbar.theme";
 @import "src/app/main/messages/messages.theme";
-@import "src/app/disclaimer/disclaimer.theme";
+@import "startup/disclaimer/disclaimer.theme";
 
 // Define a mixin for easier access
 @mixin components-theme($theme) {
