--- web/angular-wallet/src/app/disclaimer/disclaimer.component.spec.ts
@@ -1,50 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DisclaimerComponent } from './disclaimer.component';
-import { MatButtonModule } from '@angular/material/button';
-import { StoreService } from 'app/store/store.service';
-import { BehaviorSubject } from 'rxjs';
-import { RouterTestingModule } from '@angular/router/testing';
-import { FuseSharedModule } from '@fuse/shared.module';
-
-// fuggit for now, waiting for shallow render
-xdescribe('DisclaimerComponent', () => {
-  let component: DisclaimerComponent;
-  let fixture: ComponentFixture<DisclaimerComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      imports: [
-        FuseSharedModule, // it doesnt like this
-        MatButtonModule,
-        RouterTestingModule.withRoutes(
-          [{path: 'disclaimer', component: DisclaimerComponent}]
-        )
-      ],
-      declarations: [ DisclaimerComponent ],
-      providers: [ 
-        {
-          provide: StoreService,
-          useFactory: () => {
-            return {
-              ready: new BehaviorSubject(true),
-              getSettings: () => Promise.resolve({language:'en'}),
-              saveSettings: () => Promise.resolve(true)
-            }
-          }
-        }
-      ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(DisclaimerComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
