+++ web/angular-wallet/src/app/login/login-guard.service.ts
@@ -28,7 +28,12 @@ export class LoginGuard implements CanActivate {
         const selectedAccount = await this.storeService.getSelectedAccount().catch(errorHandler);
         const allAccounts = await this.storeService.getAllAccounts().catch(errorHandler);
         // caches the current nodes api version
-        await this.apiService.fetchBrsApiVersion();
+        try {
+          await this.apiService.fetchBrsApiVersion();
+        } catch (e) {
+          await this.router.navigate(['/repair']);
+          return false;
+        }
 
         // User must agree to disclaimer
         if (!(settings && settings.agree)) {
