+++ web/angular-wallet/src/app/startup/disclaimer/disclaimer.component.html
@@ -1,12 +1,12 @@
 
-<div id="disclaimer" fxLayout="column">
+<div id="repair" fxLayout="column">
 
-  <div id="disclaimer-form-wrapper" fxLayout="column" fxLayoutAlign="center center">
+  <div id="repair-form-wrapper" fxLayout="column" fxLayoutAlign="center center">
 
-      <div id="disclaimer-form" [@animate]="{value:'*',params:{duration:'300ms',y:'100px'}}">
+      <div id="repair-form" [@animate]="{value:'*',params:{duration:'300ms',y:'100px'}}">
 
           <div class="logo">
-              <img src="assets/images/logos/logo_blue.png">
+              <img src="../../../assets/images/logos/logo_blue.png">
           </div>
 
           <div class="title">Disclaimer</div>
