+++ web/angular-wallet/src/app/startup/disclaimer/disclaimer.component.ts
