+++ web/angular-wallet/src/app/startup/disclaimer/disclaimer.module.ts
@@ -1,11 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-
 import { FuseSharedModule } from '@fuse/shared.module';
 import { DisclaimerComponent } from './disclaimer.component';
-import { StoreService } from 'app/store/store.service';
 import { MatButtonModule } from '@angular/material/button';
-
 
 const routes = [
   {
