+++ web/angular-wallet/src/app/startup/repair/repair.component.html
@@ -1,0 +1,11 @@
+<div id="repair" fxLayout="column">
+  <div id="repair-form-wrapper" fxLayout="column" fxLayoutAlign="center center">
+    <div id="repair-form">
+      <div class="logo">
+        <img src="../../../assets/images/logos/logo_blue.png">
+      </div>
+      <div class="title">Repair</div>
+      <app-settings-form [settings]="settings"></app-settings-form>
+    </div>
+  </div>
+</div>
